Search Video Functionality on VideoFragment: Created menu file and implemented the working of the searchView

+++ app/src/main/java/app/netlify/dev4rju9/videoVerse/PlayerActivity.java
@@ -278,7 +278,7 @@ public class PlayerActivity extends AppCompatActivity {
                                     moveTaskToBack(true);
                                     System.exit(1);
                                 }
-                            }, (long)(sleep.get() * 60 * 1000L));
+                            }, sleep.get() * 60 * 1000L);
                             self.dismiss();
                         })
                         .setBackground(new ColorDrawable(0xB300BEF7)).create();
